refactor(toolchain): separate gcc/clang registration from definition to prevent redundant loading
