+++ xmake/toolchains/clang/xmake.lua
@@ -18,73 +18,6 @@
 -- @file        xmake.lua
 --
 
--- define toolchain
-function toolchain_clang(version)
-local suffix = ""
-if version then
-    suffix = suffix .. "-" .. version
-end
-toolchain("clang" .. suffix)
-    set_kind("standalone")
-    set_homepage("https://clang.llvm.org/")
-    set_description("A C language family frontend for LLVM" .. (version and (" (" .. version .. ")") or ""))
-    set_runtimes("c++_static", "c++_shared", "stdc++_static", "stdc++_shared")
+includes(path.join(os.scriptdir(), "toolchain_clang.lua"))
 
-    set_toolset("cc",      "clang" .. suffix)
-    set_toolset("cxx",     "clang++" .. suffix, "clang" .. suffix)
-    set_toolset("ld",      "clang++" .. suffix, "clang" .. suffix)
-    set_toolset("sh",      "clang++" .. suffix, "clang" .. suffix)
-    set_toolset("ar",      "llvm-ar" .. suffix, "ar")
-    set_toolset("strip",   "llvm-strip" .. suffix, "strip")
-    set_toolset("ranlib",  "llvm-ranlib" .. suffix, "ranlib")
-    set_toolset("objcopy", "llvm-objcopy" .. suffix, "objcopy")
-    set_toolset("nm",      "llvm-nm" .. suffix, "nm")
-    set_toolset("mm",      "clang" .. suffix)
-    set_toolset("mxx",     "clang++" .. suffix, "clang" .. suffix)
-    set_toolset("as",      "clang" .. suffix)
-    set_toolset("mrc",     "llvm-rc" .. suffix)
-    set_toolset("dlltool", "llvm-dlltool" .. suffix)
-
-    on_check(function (toolchain)
-        if toolchain:is_plat("windows") then
-            local rootdir = path.join(path.directory(os.scriptdir()), "clang")
-            local result = import("check", {rootdir = rootdir})(toolchain, suffix)
-            if result then
-                return result
-            end
-        end
-
-        return import("lib.detect.find_tool")("clang", {program = "clang" .. suffix})
-    end)
-
-    on_load(function (toolchain)
-        import("core.project.project")
-
-        if project.policy("build.optimization.lto") then
-            toolchain:set("toolset", "ar",  "llvm-ar" .. suffix)
-            toolchain:set("toolset", "ranlib",  "llvm-ranlib" .. suffix)
-        end
-
-        local march
-        if toolchain:is_arch("x86_64", "x64") then
-            march = "-m64"
-        elseif toolchain:is_arch("i386", "x86") then
-            march = "-m32"
-        end
-        if march then
-            toolchain:add("cxflags", march)
-            toolchain:add("mxflags", march)
-            toolchain:add("asflags", march)
-            toolchain:add("ldflags", march)
-            toolchain:add("shflags", march)
-        end
-        if toolchain:is_plat("windows") then
-            toolchain:add("runtimes", "MT", "MTd", "MD", "MDd")
-        end
-        if toolchain:is_plat("windows", "mingw") then
-            local rootdir = path.join(path.directory(os.scriptdir()), "clang")
-            import("load", {rootdir = rootdir})(toolchain, suffix)
-        end
-    end)
-end
 toolchain_clang()
