+++ xmake/toolchains/gcc-13/xmake.lua
@@ -1,3 +1,3 @@
-includes(path.join(os.scriptdir(), "../gcc/xmake.lua"))
+includes(path.join(os.scriptdir(), "../gcc/toolchain_gcc.lua"))
 
 toolchain_gcc("13")
